refactor: rewrite user molecule in typescript
